challanges/web/Cryptweb/src/requirements.txt
Flask==2.3.2
pycryptodome==3.20.0